+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## WIP
+- Update Electron to 8.2.3
 - Add button to shuffle/repeat songs
 - Add ability to sort tables
 - Fix sidebar z-index